Moves "onUnhandledRequest" into "utils/request" directory
