+++ src/node/setupServer.ts
@@ -11,7 +11,7 @@ import { parseBody } from '../utils/request/parseBody'
 import { isNodeProcess } from '../utils/internal/isNodeProcess'
 import * as requestHandlerUtils from '../utils/handlers/requestHandlerUtils'
 import { SharedOptions } from '../sharedOptions'
-import { onUnhandledRequest } from '../onUnhandledRequest'
+import { onUnhandledRequest } from '../utils/request/onUnhandledRequest'
 
 type ListenOptions = SharedOptions
